requirements.txt
attrs==21.4.0
bcrypt==3.2.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
flask-mongoengine==1.0.0
flask-talisman==1.0.0
Flask-WTF==1.0.1
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
